requirements.txt
. .venv/bin/Activate.ps1google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.14.0
google-cloud-datacatalog>=3.16.0
google-cloud-secret-manager>=2.16.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyGithub>=2.1.1
google-cloud-discoveryengine>=0.16.0
google-cloud-bigquery>=3.10.0
google-cloud-dataplex>=1.10.0
# Siguiente iteración
# Flask>=3.0.0
# gunicorn>=21.2.0